moving code from api to viewer folder (this folder will contain code only related to testing graphs,
