+++ plot-editor/viewer/viewerController.js
